[refact] route 경로, 페이지 수정
