+++ frontend/src/layout/index.tsx
@@ -1,22 +1,38 @@
 import React from 'react';
-import { Outlet } from 'react-router-dom';
 import { ThemeProvider } from 'styled-components';
+import { Col, Row } from 'react-bootstrap';
 
-import Header from '@/components/common/Header';
-import Footer from '@/components/common/Footer';
+import Header from '@components/common/Header';
+import Footer from '@components/common/Footer';
+import AdminLinks from '@components/AdminLinks';
+import UserChat from '@components/chat/UserChat';
 
 // CSS
 import GlobalStyles from '@styles/GlobalStyles';
 import theme from '@styles/theme';
 import * as S from './style';
 
-const Layout = () => {
+interface LayoutProps {
+  children: React.ReactNode;
+  adminLinks?: boolean;
+}
+
+const Layout = ({ children, adminLinks }: LayoutProps) => {
   return (
     <ThemeProvider theme={theme}>
       <Header />
-      <S.LayoutWrapper>
-        <Outlet />
-      </S.LayoutWrapper>
+      {adminLinks ? (
+        <S.LayoutWrapper>
+          <Row className="m-5">
+            <Col md={2}>
+              <AdminLinks />
+            </Col>
+            <Col md={10}>{children}</Col>
+          </Row>
+        </S.LayoutWrapper>
+      ) : (
+        <S.LayoutWrapper>{children}</S.LayoutWrapper>
+      )}
       <Footer />
       <GlobalStyles />
     </ThemeProvider>
