+++ frontend/src/pages/admin/Chat/index.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import { Row } from 'react-bootstrap';
+
+import Layout from '@layout/index';
+import AdminChat from '@components/chat/AdminChat';
+
+const AdminChatsPage = () => {
+  return (
+    <Layout adminLinks>
+      <Row>
+        <AdminChat />
+      </Row>
+    </Layout>
+  );
+};
+
+export default AdminChatsPage;
