+++ frontend/src/pages/admin/EditProduct/index.tsx
@@ -1,8 +1,26 @@
 import React, { useState } from 'react';
-import { Alert, Button, CloseButton, Col, Container, Form, Row, Table } from 'react-bootstrap';
+import {
+  Alert,
+  Button,
+  CloseButton,
+  Col,
+  Container,
+  Form,
+  Image,
+  Row,
+  Table,
+} from 'react-bootstrap';
 import { Link } from 'react-router-dom';
 
-const AdminCreateProductPage = () => {
+const onHover: React.CSSProperties = {
+  cursor: 'pointer',
+  position: 'absolute',
+  left: '5px',
+  top: '-10px',
+  transform: 'scale(2.7)',
+};
+
+const EditProduct = () => {
   const [validated, setValidated] = useState(false);
 
   const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
@@ -25,50 +43,47 @@ const AdminCreateProductPage = () => {
           </Link>
         </Col>
         <Col md={6}>
-          <h1>상품 추가</h1>
+          <h1>상품 수정</h1>
           <Form noValidate validated={validated} onSubmit={handleSubmit}>
             {/* 상품명 */}
             <Form.Group className="mb-3" controlId="formBasicName">
               <Form.Label>상품명</Form.Label>
-              <Form.Control name="name" required type="text" />
+              <Form.Control name="name" required type="text" defaultValue="상품명" />
             </Form.Group>
 
             {/* 상품 설명 */}
             <Form.Group className="mb-3" controlId="exampleForm.ControlTextarea1">
               <Form.Label>상품 설명</Form.Label>
-              <Form.Control name="description" required as="textarea" rows={3} />
+              <Form.Control
+                name="description"
+                required
+                as="textarea"
+                rows={3}
+                defaultValue="상품 설명"
+              />
             </Form.Group>
 
             {/* 재고 */}
             <Form.Group className="mb-3" controlId="formBasicCount">
               <Form.Label>재고</Form.Label>
-              <Form.Control name="count" required type="number" />
+              <Form.Control name="count" required type="number" defaultValue="20" />
             </Form.Group>
 
             {/* 가격 */}
             <Form.Group className="mb-3" controlId="formBasicPrice">
               <Form.Label>가격</Form.Label>
-              <Form.Control name="price" required type="text" />
+              <Form.Control name="price" required type="text" defaultValue="100000" />
             </Form.Group>
 
             {/* 카테고리 */}
             <Form.Group className="mb-3" controlId="formBasicCategory">
-              <Form.Label>
-                카테고리
-                <CloseButton />(<small>선택 지우기</small>)
-              </Form.Label>
+              <Form.Label>카테고리</Form.Label>
               <Form.Select id="cats" required name="category" aria-label="Default select example">
                 <option value="">카테고리 선택</option>
                 <option value="1">노트북</option>
                 <option value="2">카메라</option>
                 <option value="3">게임 타이틀</option>
               </Form.Select>
-            </Form.Group>
-
-            {/* 새 카테고리 추가 */}
-            <Form.Group className="mb-3" controlId="formBasicNewCategory">
-              <Form.Label>새 카테고리 추가</Form.Label>
-              <Form.Control name="newCategory" type="text" />
             </Form.Group>
 
             {/* 속성 */}
@@ -134,11 +149,21 @@ const AdminCreateProductPage = () => {
             {/* 이미지 */}
             <Form.Group controlId="formFileMultiple" className="mb-3 mt-3">
               <Form.Label>상품 이미지</Form.Label>
+              <Row>
+                <Col style={{ position: 'relative' }} xs={3}>
+                  <Image crossOrigin="anonymous" src="/images/games-category.png" fluid />
+                  <i style={onHover} className="bi bi-x text-danger" />
+                </Col>
+                <Col style={{ position: 'relative' }} xs={3}>
+                  <Image crossOrigin="anonymous" src="/images/games-category.png" fluid />
+                  <i style={onHover} className="bi bi-x text-danger" />
+                </Col>
+              </Row>
               <Form.Control required name="images" type="file" multiple />
             </Form.Group>
 
             <Button variant="primary" type="submit">
-              상품 등록
+              상품 업데이트
             </Button>
           </Form>
         </Col>
@@ -147,4 +172,4 @@ const AdminCreateProductPage = () => {
   );
 };
 
-export default AdminCreateProductPage;
+export default EditProduct;
