+++ frontend/src/pages/admin/index.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import { Routes } from 'react-router-dom';
+
+import { generateRoutes } from '@/utils/routes/generateRoutes';
+import { adminProtectedRoutes } from '@/routes/adminProtectedRoutes';
+
+const AdminPage = () => {
+  return (
+    <>
+      <Routes>{generateRoutes(adminProtectedRoutes)}</Routes>
+    </>
+  );
+};
+
+export default AdminPage;
