+++ frontend/src/utils/routes/UserChatRoutes.tsx
@@ -6,7 +6,6 @@ import UserChat from '@/components/chat/UserChat';
 const UserChatRoutes = () => {
   return (
     <>
-      <UserChat />
       <Outlet />
     </>
   );
